plugins: use yaml-config instead of defines
This is more useful during the development, because it is not neccessary
to change and recompile the source file when just changing a config
value.

+++ plugins/src/plugins/mps/ring_station.cpp
@@ -124,7 +124,7 @@ void RingStation::new_machine_info(ConstMachine &machine)
   
   // show number of bases
   number_bases_ = machine.loaded_with();
-  for(u_int32_t i=0; i < MAX_NUM_BASES; i++)
+  for(u_int32_t i=0; i < (u_int32_t) MAX_NUM_BASES; i++)
   {
     publish_indicator(i < machine.loaded_with(), i);
   }
